requirements.txt
streamlit
torch
Pillow
transformers
chromadb
langchain-core
langchain-community